// src/API.ts
export type CreateBlogInput = {
  id?: string | null,
  name: string,
};
// ...
export type CreatePostInput = {
  id?: string | null,
  title: string,
  blogPostsId?: string | null,
};
// ...
export type CreateCommentInput = {
  id?: string | null,
  content: string,
  postCommentsId?: string | null,
};